+++ bin/configurations/application.js
@@ -5,6 +5,11 @@ const os = require('os')
 
 class ApplicationConfiguration {
   constructor () {
+    this.__paths = {
+      application: `configuration/application.json`,
+      global: `${os.homedir()}/.vaultrc`,
+      local: `${process.cwd()}/.vaultrc`
+    }
     this.__configuration = {}
   }
 
@@ -12,11 +17,13 @@ class ApplicationConfiguration {
     this.__configuration = _.merge({}, this.__configuration, manifestJson)
   }
 
-  load () {
+  initialize () {
+    const paths = this.__paths
+
     _.each([
-      `configuration/application.json`,
-      `${os.homedir()}/.vaultrc`,
-      `${process.cwd()}/.vaultrc`
+      paths.application,
+      paths.global,
+      paths.local
     ], (configurationPath) => {
       this.loadFile(configurationPath)
     })
@@ -24,12 +31,35 @@ class ApplicationConfiguration {
     return this
   }
 
+  saveGlobalFile (setting) {
+    return this.saveFile(`global`, settings)
+  }
+
+  saveLocalFile (settings) {
+    return this.saveFile(`local`, settings)
+  }
+
+  saveFile(type, settings) {
+    const filename = this.__paths[type]
+    const configurationContent = this.__getFile(filename)
+
+    return this._saveFile(filename, _.merge({}, configurationContent, settings))
+  }
+
+  _saveFile (filename, content) {
+    fs.writeFileSync(filename, JSON.stringify(content, null, 2))
+    return this
+  }
+
   loadFile (filename) {
+    this.__configuration = _.merge(this.__configuration, this.__getFile(filename))
+    return this
+  }
+
+  __getFile (filename) {
     try {
-      this.__configuration = _.merge(this.__configuration,
-        JSON.parse(
-          fs.readFileSync(filename).toString() || '{}'
-        )
+      return JSON.parse(
+        fs.readFileSync(filename).toString() || '{}'
       )
     } catch (e) {
       if (e instanceof SyntaxError) {
@@ -45,4 +75,4 @@ class ApplicationConfiguration {
   }
 }
 
-module.exports = new ApplicationConfiguration().load()
+module.exports = new ApplicationConfiguration().initialize()