+++ bin/parsers/reference_parser.js
@@ -6,7 +6,7 @@ const applicationConfiguration = require('../configurations/application')
 
 const Discover = {
   COMPONENT_ASPECT: /.*?\/([a-z0-9-_.]*?)[./]?(zip|tar|tgz|gz|tar.gz|git)?$/i,
-  HAS_EMBEDDED_VERSION: /[]*?[_-]+([\d.]*)$/i
+  HAS_EMBEDDED_VERSION: /.*?[_-]+([\d.]*)$/i
 }
 
 /**
@@ -19,7 +19,7 @@ const Discover = {
  * "uri#version" ===> { url, version }
  */
 class ReferenceParser {
-  static referenceToArchiveRequest(reference, overrideUniqueName = undefined) {
+  static referenceToArchiveRequest (reference, overrideUniqueName = undefined) {
     const scopeOrReference = this.normalizeReference(reference)
     const {
       resource,
@@ -39,16 +39,16 @@ class ReferenceParser {
    * @param { reference } reference
    * @returns "reference"
    */
-  static normalizeReference(reference) {
+  static normalizeReference (reference) {
     return _.trimEnd(_.trimStart(reference || '', path.sep))
   }
 
-  static splitArchiveExtension(uri) {
+  static splitArchiveExtension (uri) {
     const [_0, archive, extension] = Discover.COMPONENT_ASPECT.exec(uri) || [``, uri, ``] // eslint-disable-line
     return [archive, extension]
   }
 
-  static splitURIVersion(reference) {
+  static splitURIVersion (reference) {
     const patternMarkers = applicationConfiguration.get(`rules.patternMarkers`)
     const versionMarker = _.find(patternMarkers.version, (versionMarker) => reference.indexOf(versionMarker) !== -1)
 
@@ -59,7 +59,7 @@ class ReferenceParser {
    *
    * @param {*} scopeOrReference
    */
-  static __scopeToResource(scopeOrReference) {
+  static __scopeToResource (scopeOrReference) {
     const patternMarkers = applicationConfiguration.get(`rules.patternMarkers`)
     const [uri, version = ``] = this.splitURIVersion(scopeOrReference)
     const uriAspects = uri.split(patternMarkers.separator)
@@ -72,10 +72,9 @@ class ReferenceParser {
       }
     }
 
-    // TODO when we do `publish` we should use `push_uri` and configure this based on the operation
     const {
       pattern,
-      pull_uri,
+      pull_uri, // OR push_uri if operation is publish
       constants
     } = scope
     const templateVariables = _.zipObject(pattern.split(patternMarkers.separator), uriAspects)
@@ -92,7 +91,7 @@ class ReferenceParser {
    *
    * @param {*} resource
    */
-  static __resourceToArchiveRequest(resource, scope, overrideUniqueName = undefined) {
+  static __resourceToArchiveRequest (resource, scope, overrideUniqueName = undefined) {
     const versionMarker = _.first(applicationConfiguration.get(`rules.patternMarkers.version`))
     const {
       staging,
@@ -101,8 +100,7 @@ class ReferenceParser {
 
     const [uri, _version] = this.splitURIVersion(resource)
     const [_archive, extension] = this.splitArchiveExtension(uri)
-    // TODO should I default to * because master is for git?
-    const [archive, version = `master`] = this.__detectVersionInArchive(_archive, _version)
+    const [archive, version = `*`] = this.__detectVersionInArchive(_archive, _version)
 
     const versionFolder = crypto.createHash(`md5`).update(version).digest(`hex`)
     const installedName = overrideUniqueName || archive
@@ -111,8 +109,6 @@ class ReferenceParser {
     const cachePath = `${cache}/${archive}__${versionFolder}${safeExtension}`
     const stagingPath = `${staging}/${archive}/${versionFolder}/`
     const uuid = `${installedName}${versionMarker}${version}`
-
-    // TODO pass unique name through to collect -v2 and v-3 and apply unique name
 
     return {
       installedVersion: version,
@@ -127,7 +123,7 @@ class ReferenceParser {
     }
   }
 
-  static __detectVersionInArchive(archive, version) {
+  static __detectVersionInArchive (archive, version) {
     const embeddedVersion = Discover.HAS_EMBEDDED_VERSION.exec(archive)
 
     if (embeddedVersion) {
@@ -137,13 +133,12 @@ class ReferenceParser {
     }
   }
 
-
   /**
    *
    * @param {*} uriAspects
    * @returns a scope object or undefined
    */
-  static __matchConversionRule(uriAspects) {
+  static __matchConversionRule (uriAspects) {
     const sources = applicationConfiguration.get(`sources`)
     const scope = uriAspects[0] = (_.first(uriAspects) || '')